// packages/internal-api/src/enums/service.type.enum.ts
export enum ServicesEnum {
  AUTH_SERVICE = 'AUTH_SERVICE',
  COMMUNICATION_SERVICE = 'COMMUNICATION_SERVICE',
  ME_SERVICE = 'ME_SERVICE',
  MESSAGE_SERVICE = 'MESSAGE_SERVICE',
  NOTIFICATION_SERVICE = 'NOTIFICATION_SERVICE',
  PAYMENT_SERVICE = 'PAYMENT_SERVICE',
  PROPERTY_SERVICE = 'PROPERTY_SERVICE',
  REVIEW_RATING_SERVICE = 'REVIEW_RATING_SERVICE',
  WHITELABEL_API = 'WHITELABEL_API',
}